requirements.txt
agno==1.6.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.38.36
botocore==1.38.36
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
curl-cffi==0.11.3
docstring-parser==0.15
duckduckgo-search==8.0.4
frozendict==2.4.6
gitdb==4.0.12
gitpython==3.1.44
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib-metadata==8.7.0
jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lxml==5.4.0
markdown-it-py==3.0.0
markupsafe==3.0.2
mcp==1.9.4
mdurl==0.1.2
multitasking==0.0.11
narwhals==1.42.1
numpy==2.3.0
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
packaging==24.2
pandas==2.3.0
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
primp==0.15.0
protobuf==5.29.5
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.6
pydantic-core==2.33.2
pydantic-settings==2.9.1
pydeck==0.9.1
pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.4
rich==14.0.0
rpds-py==0.25.1
s3transfer==0.13.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
sse-starlette==2.3.6
starlette==0.47.0
strands-agents==0.1.7
streamlit==1.45.1
tenacity==9.1.2
toml==0.10.2
tomli==2.2.1
tornado==6.5.1
typer==0.16.0
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
watchdog==6.0.0
websockets==15.0.1
yfinance==0.2.63
zipp==3.23.0
